refactor: move thread header to new component
